Split source DB maintenance job
